Use project name (eventbus) as artifact ID
To simplify buildscripts and to prevent needing to change multiple
values if need be, we've begun naming projects based on their artifact
names and using that in the maven publications instead of manually
writing out the desired artifact name (potentially multiple times).

This required changing the project name from "EventBus" to "eventbus".