Enhance alwaysCancelling listener awareness in invoker optimisation strategy (#82)
If the event is cancellable but all listeners are consumers, the invoker optimisation strategy would previously bail out to using the wrapped predicates if alwaysCancelling listeners were present.

Now, alwaysCancelling listeners are properly accounted for in this case, combining the benefits of eliminating non-monitoring listeners after them with also unwrapping to branchless consumer calls in the built invoker.